Updated registry autodiscovery
